# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI CAD Preview API",
    "description": "CadQuery primitives exposed through FastAPI and OpenAPI.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models": {
      "get": {
        "summary": "List Models",
        "operationId": "list_models_api_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalogResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{shape_type}": {
      "get": {
        "summary": "Read Model",
        "operationId": "read_model_api_models__shape_type__get",
        "parameters": [
          {
            "name": "shape_type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "box",
                "sphere",
                "cylinder"
              ],
              "type": "string",
              "description": "Primitive model type to preview.",
              "title": "Shape Type"
            },
            "description": "Primitive model type to preview."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BoxPreview"
                    },
                    {
                      "$ref": "#/components/schemas/SpherePreview"
                    },
                    {
                      "$ref": "#/components/schemas/CylinderPreview"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "shape_type",
                    "mapping": {
                      "box": "#/components/schemas/BoxPreview",
                      "sphere": "#/components/schemas/SpherePreview",
                      "cylinder": "#/components/schemas/CylinderPreview"
                    }
                  },
                  "title": "Response Read Model Api Models  Shape Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/preview": {
      "post": {
        "summary": "Preview Model",
        "operationId": "preview_model_api_models_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/BoxPreviewRequest"
                  },
                  {
                    "$ref": "#/components/schemas/SpherePreviewRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CylinderPreviewRequest"
                  }
                ],
                "title": "Request",
                "description": "Primitive parameters to rebuild with CadQuery."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BoxPreview"
                    },
                    {
                      "$ref": "#/components/schemas/SpherePreview"
                    },
                    {
                      "$ref": "#/components/schemas/CylinderPreview"
                    }
                  ],
                  "title": "Response Preview Model Api Models Preview Post",
                  "discriminator": {
                    "propertyName": "shape_type",
                    "mapping": {
                      "box": "#/components/schemas/BoxPreview",
                      "sphere": "#/components/schemas/SpherePreview",
                      "cylinder": "#/components/schemas/CylinderPreview"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BoxPreview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "center": {
            "$ref": "#/components/schemas/Vector3"
          },
          "volume": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/Vector3"
          },
          "shape_type": {
            "type": "string",
            "enum": [
              "box"
            ],
            "const": "box",
            "title": "Shape Type",
            "default": "box"
          },
          "width": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Height"
          },
          "depth": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Depth"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "color",
          "center",
          "volume",
          "bounding_box",
          "width",
          "height",
          "depth"
        ],
        "title": "BoxPreview"
      },
      "BoxPreviewRequest": {
        "properties": {
          "center": {
            "$ref": "#/components/schemas/Vector3"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "shape_type": {
            "type": "string",
            "enum": [
              "box"
            ],
            "const": "box",
            "title": "Shape Type",
            "default": "box"
          },
          "width": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Height"
          },
          "depth": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Depth"
          }
        },
        "type": "object",
        "required": [
          "center",
          "width",
          "height",
          "depth"
        ],
        "title": "BoxPreviewRequest"
      },
      "CylinderPreview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "center": {
            "$ref": "#/components/schemas/Vector3"
          },
          "volume": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/Vector3"
          },
          "shape_type": {
            "type": "string",
            "enum": [
              "cylinder"
            ],
            "const": "cylinder",
            "title": "Shape Type",
            "default": "cylinder"
          },
          "radius": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Radius"
          },
          "height": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "color",
          "center",
          "volume",
          "bounding_box",
          "radius",
          "height"
        ],
        "title": "CylinderPreview"
      },
      "CylinderPreviewRequest": {
        "properties": {
          "center": {
            "$ref": "#/components/schemas/Vector3"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "shape_type": {
            "type": "string",
            "enum": [
              "cylinder"
            ],
            "const": "cylinder",
            "title": "Shape Type",
            "default": "cylinder"
          },
          "radius": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Radius"
          },
          "height": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "center",
          "radius",
          "height"
        ],
        "title": "CylinderPreviewRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok"
            ],
            "const": "ok",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "ModelCatalogResponse": {
        "properties": {
          "models": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BoxPreview"
                },
                {
                  "$ref": "#/components/schemas/SpherePreview"
                },
                {
                  "$ref": "#/components/schemas/CylinderPreview"
                }
              ],
              "discriminator": {
                "propertyName": "shape_type",
                "mapping": {
                  "box": "#/components/schemas/BoxPreview",
                  "cylinder": "#/components/schemas/CylinderPreview",
                  "sphere": "#/components/schemas/SpherePreview"
                }
              }
            },
            "type": "array",
            "title": "Models"
          }
        },
        "type": "object",
        "required": [
          "models"
        ],
        "title": "ModelCatalogResponse"
      },
      "SpherePreview": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "center": {
            "$ref": "#/components/schemas/Vector3"
          },
          "volume": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume"
          },
          "bounding_box": {
            "$ref": "#/components/schemas/Vector3"
          },
          "shape_type": {
            "type": "string",
            "enum": [
              "sphere"
            ],
            "const": "sphere",
            "title": "Shape Type",
            "default": "sphere"
          },
          "radius": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Radius"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "color",
          "center",
          "volume",
          "bounding_box",
          "radius"
        ],
        "title": "SpherePreview"
      },
      "SpherePreviewRequest": {
        "properties": {
          "center": {
            "$ref": "#/components/schemas/Vector3"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "shape_type": {
            "type": "string",
            "enum": [
              "sphere"
            ],
            "const": "sphere",
            "title": "Shape Type",
            "default": "sphere"
          },
          "radius": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Radius"
          }
        },
        "type": "object",
        "required": [
          "center",
          "radius"
        ],
        "title": "SpherePreviewRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Vector3": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X"
          },
          "y": {
            "type": "number",
            "title": "Y"
          },
          "z": {
            "type": "number",
            "title": "Z"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "z"
        ],
        "title": "Vector3"
      }
    }
  }
}